initial speed up with boxing

+++ src/LossFunctions.jl
@@ -1,4 +1,4 @@
-function mean_squared_loss(y, ŷ, n)
+function mean_squared_loss(y, ŷ, n::Int64)
     return sum( (y .- ŷ) .^ Constant(2) ) ./ Constant(n)
 end
